fix: build errors, moved components to the components folder
