+++ components/themes/VaporWave-/index.tsx
@@ -8,7 +8,7 @@ export const getMaxPerLine = (size: number) => Math.floor(3 * Math.sqrt(size));
 
 export const splitText = (text: string) => text.split(' ').reduce((acc: string[], c: string, i, arr) => {
   // this is required as we can't just let the system automatically put the text on a new line sadly
-  const maxPerLine = getMaxPerLine(arr.join(' ').split('').length);
+  const maxPerLine = getMaxPerLine(arr.join(' ').length);
 
   while (c.length > maxPerLine) {
     console.log(`c: ${c} - too long, splitting`);
