+++ tsconfig.json
@@ -27,6 +27,6 @@
         "next-auth.d.ts",
         "**/*.ts",
         "**/*.tsx"
-    ],
+, "next.config.js"    ],
     "exclude": ["node_modules"]
 }
